Started minor changes to the coordinate system used by the touchscreen API as discussed on the mailing list.  Also beginning to separate the 'Pressed' and 'Moved' event handling on the private level.

+++ src/touchscreen/SDL_systouchscreen.h
@@ -35,9 +35,9 @@ struct _SDL_Touchscreen
     int npoints;                /* Number of points currently touched */
     struct touchpoint
     {
-        int x;
-        int y;
-        int pressure;
+        Uint16 x;
+        Uint16 y;
+        Uint16 pressure;
     } *points;                   /* Current ball motion deltas */
 
     struct touchscreen_hwdata *hwdata;     /* Driver dependent information */
